Update docs, compose config, and service controllers after worker refactor
