+++ backend-service/package.json
@@ -6,7 +6,6 @@
   "type": "module",
   "scripts": {
     "dev": "nodemon --watch src --ext ts --exec \"node --loader ts-node/esm src/index.ts\"",
-    "build:worker": "nodemon --watch src --ext ts --exec \"node --loader ts-node/esm src/job/workers/build.ts\"",
     "build": "tsc -p tsconfig.json",
     "start": "node dist/index.js"
   },
