+++ docker-compose.dev.yml
@@ -20,7 +20,6 @@ services:
     container_name: submission-service
     volumes:
       - ./submission-service:/src
-      - submission_node_modules:/src/node_modules
     ports:
       - "8002:3000"
     env_file:
@@ -36,7 +35,6 @@ services:
     container_name: extractor-service
     volumes:
       - ./extractor-service:/src
-      - extractor_node_modules:/src/node_modules
       - clones-data:/src/src/clones
       - ./extractor-service/clones:/src/src/clones
     ports:
@@ -56,7 +54,6 @@ services:
     container_name: routing-service
     volumes:
       - ./routing-service:/src
-      - routing_node_modules:/src/node_modules
       - clones-data:/src/src/clones
     ports:
       - "8004:3000"
@@ -73,13 +70,12 @@ services:
     container_name: backend-service
     volumes:
       - ./backend-service:/src
-      - backend_node_modules:/src/node_modules
       - clones-data:/src/src/clones
     ports:
       - "8005:3000"
     env_file:
       - ./backend-service/.env
-    command: [ "npm", "run", "dev" ]
+    command: [ "npm", "run", "dev"]
     networks:
       - veren
     depends_on:
@@ -87,16 +83,15 @@ services:
 
   clone-worker:
     build:
-      context: ./extractor-service
+      context: ./WORKERS/clone-worker
       dockerfile: Dockerfile.dev
     container_name: clone-worker
     command: [ "npm", "run", "clone:worker" ]
     volumes:
-      - ./extractor-service:/src
-      - extractor_node_modules:/src/node_modules
+      - ./WORKERS/clone-worker:/src
       - clones-data:/src/src/clones
     env_file:
-      - ./extractor-service/.env
+      - ./WORKERS/clone-worker/.env
     networks:
       - veren
     depends_on:
@@ -104,17 +99,16 @@ services:
 
   build-worker:
     build:
-      context: ./build-worker
+      context: ./WORKERS/build-worker
       dockerfile: Dockerfile
     privileged: true
     container_name: build-worker
     volumes:
+      - ./WORKERS/build-worker:/src
       - /var/run/docker.sock:/var/run/docker.sock
       - clones-data:/src/src/clones
-      - backend_node_modules:/src/node_modules
-      - ./backend-service:/src
     env_file:
-      - ./backend-service/.env
+      - ./WORKERS//build-worker/.env
     networks:
       - veren
     depends_on:
@@ -128,7 +122,6 @@ services:
   #   command: [ "npm", "run", "assign:worker" ]
   #   volumes:
   #     - ./extractor-service:/src
-  #     - extractor_node_modules:/src/node_modules
   #     - clones-data:/src/src/clones
   #     - ./extractor-service/src/services/assignFile/public:/src/src/services/assignFile/public
   #   env_file:
@@ -160,17 +153,10 @@ services:
     - redis-data:/data
 
 volumes:
-  api_node_modules:
-  submission_node_modules:
-  extractor_node_modules:
-  routing_node_modules:
-  backend_node_modules:
   pgdata:
   dind-storage:
   clones-data:
   redis-data:
-
-
 
 networks:
   veren:
