+++ extractor-service/package.json
@@ -6,7 +6,6 @@
   "type": "module",
   "scripts": {
     "dev": "nodemon --watch src --ext ts --exec \"node --loader ts-node/esm src/index.ts\"",
-    "clone:worker": "nodemon --watch src --ext ts --exec \"node --loader ts-node/esm src/job/workers/clone.ts\"",
     "build": "tsc -p tsconfig.json",
     "start": "node dist/index.js"
   },
